+++ data-raw/fakesurvey.R
@@ -1,0 +1,3 @@
+rm(list=ls())
+fakesurvey <- read.csv("data-raw/fakesurvey.csv")
+save(fakesurvey, file = "data/fakesurvey.rdata")